+++ Firebase/DynamicLinks/FIRDynamicLinkNetworking.m
@@ -198,8 +198,9 @@ NSData *_Nullable FIRDataWithDictionary(NSDictionary *dictionary, NSError **_Nul
   FIRDLNetworkingParserBlock responseParserBlock = ^NSDictionary *_Nullable(
       NSString *requestURLString, NSData *data, NSString **matchMessagePtr, NSError **errorPtr) {
     NSError *serializationError;
-    NSDictionary *result =
-        [NSJSONSerialization JSONObjectWithData:data options:0 error:&serializationError];
+    NSDictionary *result = [NSJSONSerialization JSONObjectWithData:data
+                                                           options:0
+                                                             error:&serializationError];
 
     if (serializationError) {
       *errorPtr = serializationError;