+++ Firestore/Example/Tests/Util/FSTIntegrationTestCase.mm
@@ -115,7 +115,6 @@ static FIRFirestoreSettings *defaultSettings;
 + (void)setUpDefaults {
   defaultSettings = [[FIRFirestoreSettings alloc] init];
   defaultSettings.persistenceEnabled = YES;
-  defaultSettings.timestampsInSnapshotsEnabled = YES;
 
   // Check for a MobileHarness configuration, running against nightly or prod, which have live
   // SSL certs.
@@ -154,11 +153,10 @@ static FIRFirestoreSettings *defaultSettings;
       [[[NSFileManager defaultManager] attributesOfItemAtPath:certsPath error:nil] fileSize];
 
   if (fileSize == 0) {
-    NSLog(
-        @"Please set up a GoogleServices-Info.plist for Firestore in Firestore/Example/App using "
-         "instructions at <https://github.com/firebase/firebase-ios-sdk#running-sample-apps>. "
-         "Alternatively, if you're a Googler with a Hexa preproduction environment, run "
-         "setup_integration_tests.py to properly configure testing SSL certificates.");
+    NSLog(@"Please set up a GoogleServices-Info.plist for Firestore in Firestore/Example/App using "
+           "instructions at <https://github.com/firebase/firebase-ios-sdk#running-sample-apps>. "
+           "Alternatively, if you're a Googler with a Hexa preproduction environment, run "
+           "setup_integration_tests.py to properly configure testing SSL certificates.");
   }
   GrpcConnection::UseTestCertificate(util::MakeString(defaultSettings.host),
                                      Path::FromNSString(certsPath), "test_cert_2");