+++ Firestore/CHANGELOG.md
@@ -1,4 +1,16 @@
 # Unreleased
+- [changed] The `areTimestampsInSnapshotsEnabled` setting is now enabled by
+  default so timestamp fields read from a FIRDocumentSnapshot will be returned
+  as FIRTimestamp objects instead of NSDate. Any code expecting to receive an
+  NSDate object must be updated.
+- [changed] `Transaction.getDocument()` has been changed to return a non-nil
+  `DocumentSnapshot` with `exists` equal to `false` if the document does not
+  exist (instead of returning a nil DocumentSnapshot). Code that was previously
+  doing `if (snapshot) { ... }` must be changed to
+  `if (snapshot.exists) { ... }`.
+- [fixed] Fixed a crash that could happen when the app is shut down after
+  a write has been sent to the server but before it has been received on
+  a listener (#2237).
 
 # v0.16.1
 - [fixed] Offline persistence now properly records schema downgrades. This is a